Changed the installation no from int to Int64

+++ src/Programs/Starcounter.UsageTracker/API/Installer/FinishCollectionHandler.POST.cs
@@ -35,7 +35,7 @@ namespace Starcounter.Applications.UsageTrackerApp.API.Installer {
                         Db.Transaction(() => {
 
                             string serial = data.downloadId;
-                            int installationNo = int.Parse(data.installationNo.ToString());
+                            Int64 installationNo = Int64.Parse(data.installationNo.ToString());
 
                             Installation installation = StarcounterCollectionHandler.AssureInstallation(installationNo, serial);
                             InstallerFinish item = new InstallerFinish(installation);
